+++ dzVents/runtime/Domoticz.lua
@@ -9,15 +9,6 @@ local TimedCommand = require('TimedCommand')
 local utils = require('Utils')
 local _ = require('lodash')
 
--- simple string splitting method
--- coz crappy LUA doesn't have this natively... *sigh*
-function string:split(sep)
-	local sep, fields = sep or ":", {}
-	local pattern = string.format("([^%s]+)", sep)
-	self:gsub(pattern, function(c) fields[#fields + 1] = c end)
-	return fields
-end
-
 -- main class
 local function Domoticz(settings)
 
@@ -124,11 +115,18 @@ local function Domoticz(settings)
 		['EVOHOME_MODE_TEMPORARY_OVERRIDE'] = 'TemporaryOverride',
 		['EVOHOME_MODE_PERMANENT_OVERRIDE'] = 'PermanentOverride',
 		['EVOHOME_MODE_FOLLOW_SCHEDULE'] = 'FollowSchedule',
+		['EVOHOME_MODE_AUTOWITHRESET'] = 'AutoWithReset',
+		['EVOHOME_MODE_AUTOWITHECO'] = 'AutoWithEco',
+		['EVOHOME_MODE_AWAY'] = 'Away',
+		['EVOHOME_MODE_DAYOFF'] = 'DayOff',
+		['EVOHOME_MODE_CUSTOM'] = 'Custom',
+		['EVOHOME_MODE_HEATINGOFF'] = 'HeatingOff',
 		['INTEGER'] = 'integer',
 		['FLOAT'] = 'float',
 		['STRING'] = 'string',
 		['DATE'] = 'date',
 		['TIME'] = 'time',
+		['NSS_FIREBASE'] = 'gcm',  -- For the moment the change to fcm is only done in the url 
 		['NSS_GOOGLE_CLOUD_MESSAGING'] = 'gcm',
 		['NSS_HTTP'] = 'http',
 		['NSS_KODI'] = 'kodi',
@@ -153,10 +151,7 @@ local function Domoticz(settings)
 			_ = _,
 
 			toCelsius = function(f, relative)
-				if (relative) then
-					return f*(1/1.8)
-				end
-				return ((f-32) / 1.8)
+				return utils.toCelsius(f, relative)
 			end,
 
 			urlEncode = function(s, strSub)
@@ -167,15 +162,8 @@ local function Domoticz(settings)
 				return utils.urlDecode(s)
 			end,
 
-			round = function(x, n)
-				n = math.pow(10, n or 0)
-				x = x * n
-				if x >= 0 then
-					x = math.floor(x + 0.5)
-				else
-					x = math.ceil(x - 0.5)
-				end
-				return x / n
+			round = function(x,n)
+				return utils.round(x,n)
 			end,
 
 			osExecute = function(cmd)
@@ -209,7 +197,10 @@ local function Domoticz(settings)
 			stringSplit = function(text, sep)
 				return utils.stringSplit(text, sep)
 			end,
-			
+
+			inTable = function(t, searchItem)
+				return utils.inTable(t, searchItem)
+			end,
 		}
 	}
 
@@ -243,6 +234,13 @@ local function Domoticz(settings)
 				_subSystem = ''
 			end
 		end
+
+		--[[
+        if _subSystem:find('gcm') then
+			utils.log('Notification subsystem Google Cloud Messaging (gcm) has been deprecated by Google. Please consider switching to Firebase', utils.LOG_ERROR)
+		end
+        ]] --
+        
 		local data = subject
 				.. '#' .. message
 				.. '#' .. tostring(priority)
@@ -385,7 +383,7 @@ local function Domoticz(settings)
 	function self.round(x, n)
 		utils.log('domoticz.round deprecated. Please use domoticz.utils.round.', utils.LOG_INFO)
 		return self.utils.round(x, n)
-	end
+	 end
 
 	function self.dump()
 		self.utils.dumpTable(settings, '> ')