+++ dzVents/runtime/Utils.lua
@@ -7,7 +7,7 @@ local self = {
 	LOG_MODULE_EXEC_INFO = 2,
 	LOG_INFO = 3,
 	LOG_DEBUG = 4,
-	DZVERSION = '2.4.19',
+	DZVERSION = '2.4.24',
 }
 
 function self.fileExists(name)
@@ -27,6 +27,35 @@ function self.stringSplit(text, sep)
 		table.insert(t, str)
 	end
 	return t
+end
+
+function self.inTable(searchTable, element)
+	if type(searchTable) ~= 'table' then return false end
+	local res = res
+	for k, v in pairs(searchTable) do
+		if type(v) == 'table' then res = self.inTable(v, element) end
+		res = res or (( tostring(k) == tostring(element) and 'key' ) or ( tostring(v) == tostring(element) and 'value' ))
+		if res then return res end
+	end
+	return false
+end
+
+function self.round(x, n)
+	n = math.pow(10, n or 0)
+	x = x * n
+	if x >= 0 then
+		x = math.floor(x + 0.5)
+	else
+		x = math.ceil(x - 0.5)
+	end
+	return x / n
+end
+
+function self.toCelsius(f, relative)
+	if (relative) then
+		return f*(1/1.8)
+	end
+	return ((f-32) / 1.8)
 end
 
 function self.osExecute(cmd)
@@ -55,6 +84,7 @@ function self.urlEncode(str, strSub)
 	end
 	return str
 end
+
 function self.urlDecode(str, strSub)
 
 	local hex2Char = function(x)
@@ -124,7 +154,7 @@ function self.log(msg, level)
 
 
 	if (level == self.LOG_ERROR) then
-		marker = marker .. 'Error (' .. self.DZVERSION .. '): '
+		marker = marker .. 'Error: (' .. self.DZVERSION .. ') '
 	elseif (level == self.LOG_DEBUG) then
 		marker = marker .. 'Debug: '
 	elseif (level == self.LOG_INFO or level == self.LOG_MODULE_EXEC_INFO) then
@@ -183,7 +213,7 @@ function self.dumpTable(t, level)
 		if (type(value) ~= 'function') then
 			if (type(value) == 'table') then
 				self.print(level .. attr .. ':')
-				self.dumpTable(value, level .. '    ')
+				self.dumpTable(value, level .. '	')
 			else
 				self.print(level .. attr .. ': ' .. tostring(value))
 			end