+++ webserver/cWebem.cpp
@@ -504,7 +504,7 @@ namespace http {
 
 			request_path = ExtractRequestPath(request_path);
 
-			int paramPos = request_path.find_first_of('?');
+			size_t paramPos = request_path.find_first_of('?');
 			if (paramPos != std::string::npos)
 			{
 				request_path = request_path.substr(0, paramPos);
@@ -703,20 +703,20 @@ namespace http {
 					}
 				}
 
+				reply::add_header(&rep, "Content-Length", std::to_string(rep.content.size()));
 				if (!boost::algorithm::starts_with(strMimeType, "image"))
 				{
-					reply::add_header(&rep, "Content-Length", std::to_string(rep.content.size()));
-					reply::add_header(&rep, "Content-Type", strMimeType + ";charset=UTF-8");
+					if (!strMimeType.empty())
+						strMimeType += ";charset=UTF-8";
 					reply::add_header(&rep, "Cache-Control", "no-cache");
 					reply::add_header(&rep, "Pragma", "no-cache");
 					reply::add_header(&rep, "Access-Control-Allow-Origin", "*");
 				}
 				else
 				{
-					reply::add_header(&rep, "Content-Length", std::to_string(rep.content.size()));
-					reply::add_header(&rep, "Content-Type", strMimeType);
 					reply::add_header(&rep, "Cache-Control", "max-age=3600, public");
 				}
+				reply::add_header_content_type(&rep, strMimeType);
 				return true;
 			}
 
@@ -994,6 +994,11 @@ namespace http {
 		const std::string cWebem::GetPort()
 		{
 			return m_settings.listening_port;
+		}
+
+		const std::string cWebem::GetWebRoot()
+		{
+			return m_webRoot;
 		}
 
 		WebEmSession * cWebem::GetSession(const std::string & ssid)
@@ -1411,7 +1416,7 @@ namespace http {
 				bool bHaveGZipSupport = (strstr(encoding_header, "gzip") != NULL);
 				if (bHaveGZipSupport)
 				{
-					CA2GZIP gzip((char*)rep.content.c_str(), rep.content.size());
+					CA2GZIP gzip((char*)rep.content.c_str(), (int)rep.content.size());
 					if ((gzip.Length > 0) && (gzip.Length < (int)rep.content.size()))
 					{
 						rep.bIsGZIP = true; // flag for later
@@ -1931,8 +1936,8 @@ namespace http {
 				if (cookie != NULL)
 				{
 					std::string scookie = cookie;
-					int fpos = scookie.find("DMZSID=");
-					int upos = scookie.find("_", fpos);
+					size_t fpos = scookie.find("DMZSID=");
+					size_t upos = scookie.find("_", fpos);
 					if ((fpos != std::string::npos) && (upos != std::string::npos))
 					{
 						std::string sSID = scookie.substr(fpos + 7, upos - fpos - 7);
@@ -1995,6 +2000,9 @@ namespace http {
 			modify_info mInfo;
 			if (myWebem->CheckForPageOverride(session, requestCopy, rep))
 			{
+				if (rep.status == reply::status_type::download_file)
+					return;
+
 				if (session.reply_status != reply::ok) // forbidden
 				{
 					rep = reply::stock_reply(static_cast<reply::status_type>(session.reply_status));
@@ -2021,11 +2029,12 @@ namespace http {
 				// do normal handling
 				try
 				{
-					if (requestCopy.uri.find("/images/") == 0)
+					std::string uri = myWebem->ExtractRequestPath(requestCopy.uri);
+					if (uri.find("/images/") == 0)
 					{
-						std::string theme_images_path = myWebem->m_actTheme + requestCopy.uri;
+						std::string theme_images_path = myWebem->m_actTheme + uri;
 						if (file_exist((doc_root_ + theme_images_path).c_str()))
-							requestCopy.uri = theme_images_path;
+							requestCopy.uri = myWebem->GetWebRoot() + theme_images_path;
 					}
 
 					request_handler::handle_request(requestCopy, rep, mInfo);